requirements.txt
fastapi
uvicorn
requests
sqlite-utils
sentence-transformers
faiss-cpu
numpy
google-generativeai
python-dotenv
